+++ Makcha/src/components/dashboard/index.tsx
@@ -2,8 +2,8 @@ import { useEffect, useCallback } from 'react';
 import { Link } from 'react-router-dom';
 import { motion } from 'framer-motion';
 import { ChevronsLeft, ChevronsRight, X } from 'lucide-react';
-import DashboardNav from './DashboardNav';
-import DarkModeToggle from './DarkModeToggle'; 
+import DashboardNav from './components/DashboardNav';
+import DarkModeToggle from './components/DarkModeToggle'; 
 import { useAuth } from '../../hooks/useAuth';
 import { sidebarSpring } from './constants';
 import { useDashboardStore } from '../../store/useDashboardStore';
