+++ Makcha/src/hooks/useGeoLocation.ts
@@ -5,33 +5,31 @@ interface ILocation {
   longitude: number;
 }
 
-export const useGeoLocation = (
-  options?: PositionOptions
-) => {
+const isGeolocationSupported = typeof window !== "undefined" && "geolocation" in navigator;
+
+export const useGeoLocation = (options?: PositionOptions) => {
   const [location, setLocation] = useState<ILocation | null>(null);
-  const [error, setError] = useState("");
-  const [loading, setLoading] = useState(true);
+
+  const [error, setError] = useState(isGeolocationSupported ? "" : "Geolocation is not supported.");
+  const [loading, setLoading] = useState(isGeolocationSupported);
 
   useEffect(() => {
-    if (!navigator.geolocation) {
-      setError("Geolocation is not supported.");
-      setLoading(false);
-      return;
-    }
 
-    navigator.geolocation.getCurrentPosition(
-      (pos) => {
-        const { latitude, longitude } = pos.coords;
-        setLocation({ latitude, longitude });
-        setLoading(false);
-      },
-      (err) => {
-        setError(err.message);
-        setLoading(false);
-      },
-      options
-    );
-  }, []);
+    if (!isGeolocationSupported) return;
+
+    const handleSuccess = (pos: GeolocationPosition) => {
+      const { latitude, longitude } = pos.coords;
+      setLocation({ latitude, longitude });
+      setLoading(false);
+    };
+
+    const handleError = (err: GeolocationPositionError) => {
+      setError(err.message);
+      setLoading(false);
+    };
+
+    navigator.geolocation.getCurrentPosition(handleSuccess, handleError, options);
+  }, [options]);
 
   return { location, error, loading };
 };
