+++ Makcha/src/pages/Download/types/download.ts
@@ -1,0 +1,4 @@
+export interface DownloadButtonProps {
+    onClick: () => void;
+    isInstallable: boolean; // 설치 가능 여부
+  }
