+++ Makcha/src/pages/Home/types/home.ts
@@ -1,0 +1,7 @@
+export interface DashboardAction {
+    id: string;
+    title: string;
+    description: string;
+    path: string;
+    variant: "primary" | "secondary";
+  }
