+++ Makcha/src/pages/Setting/index.tsx
@@ -3,8 +3,8 @@ import { useLocation, useNavigate } from "react-router-dom";
 import { SettingPanel } from "./components/SettingPanel";
 import { SettingBg } from "./components/BgSection";
 import { useSettingStore } from "../../store/useSettingStore";
-import { useBack } from "../../hooks/useBack";
-import type { Place } from "../../types/setting";
+import { useBack } from "./hooks/useBack";
+import type { Place } from "./types/setting";
 import type { ViewType } from "./constants";
 import PlaceSetting from "./components/PlaceSetting";
 import { PhonenumberSetting } from "./components/PhonenumberSetting";
